Create SystemIndexRegistry
