+++ server/src/test/java/org/opensearch/cluster/ClusterModuleTests.java
@@ -70,7 +70,6 @@ import org.opensearch.common.settings.Settings;
 import org.opensearch.common.settings.SettingsModule;
 import org.opensearch.common.util.concurrent.ThreadContext;
 import org.opensearch.gateway.GatewayAllocator;
-import org.opensearch.indices.SystemIndices;
 import org.opensearch.plugins.ClusterPlugin;
 import org.opensearch.telemetry.metrics.noop.NoopMetricsRegistry;
 import org.opensearch.test.ClusterServiceUtils;
@@ -169,13 +168,7 @@ public class ClusterModuleTests extends ModuleTestCase {
                 public Collection<AllocationDecider> createAllocationDeciders(Settings settings, ClusterSettings clusterSettings) {
                     return Collections.singletonList(new EnableAllocationDecider(settings, clusterSettings));
                 }
-            }),
-                clusterInfoService,
-                null,
-                threadContext,
-                new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE),
-                new SystemIndices(Collections.emptyMap())
-            )
+            }), clusterInfoService, null, threadContext, new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE))
         );
         assertEquals(e.getMessage(), "Cannot specify allocation decider [" + EnableAllocationDecider.class.getName() + "] twice");
     }
@@ -186,13 +179,7 @@ public class ClusterModuleTests extends ModuleTestCase {
             public Collection<AllocationDecider> createAllocationDeciders(Settings settings, ClusterSettings clusterSettings) {
                 return Collections.singletonList(new FakeAllocationDecider());
             }
-        }),
-            clusterInfoService,
-            null,
-            threadContext,
-            new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE),
-            new SystemIndices(Collections.emptyMap())
-        );
+        }), clusterInfoService, null, threadContext, new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE));
         assertTrue(module.deciderList.stream().anyMatch(d -> d.getClass().equals(FakeAllocationDecider.class)));
     }
 
@@ -202,13 +189,7 @@ public class ClusterModuleTests extends ModuleTestCase {
             public Map<String, Supplier<ShardsAllocator>> getShardsAllocators(Settings settings, ClusterSettings clusterSettings) {
                 return Collections.singletonMap(name, supplier);
             }
-        }),
-            clusterInfoService,
-            null,
-            threadContext,
-            new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE),
-            new SystemIndices(Collections.emptyMap())
-        );
+        }), clusterInfoService, null, threadContext, new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE));
     }
 
     public void testRegisterShardsAllocator() {
@@ -236,8 +217,7 @@ public class ClusterModuleTests extends ModuleTestCase {
                 clusterInfoService,
                 null,
                 threadContext,
-                new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE),
-                new SystemIndices(Collections.emptyMap())
+                new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE)
             )
         );
         assertEquals("Unknown ShardsAllocator [dne]", e.getMessage());
@@ -323,8 +303,7 @@ public class ClusterModuleTests extends ModuleTestCase {
             clusterInfoService,
             null,
             threadContext,
-            new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE),
-            new SystemIndices(Collections.emptyMap())
+            new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE)
         );
         expectThrows(
             IllegalArgumentException.class,
@@ -340,8 +319,7 @@ public class ClusterModuleTests extends ModuleTestCase {
             clusterInfoService,
             null,
             threadContext,
-            new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE),
-            new SystemIndices(Collections.emptyMap())
+            new ClusterManagerMetrics(NoopMetricsRegistry.INSTANCE)
         );
         expectThrows(
             IllegalArgumentException.class,
